+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2023-02-24
+
 ### Breaking
 
 - Refactored `GameInterface` into a struct that is generic over any backend-implementation
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Release
 
-[unreleased]: https://github.com/BfBBModdingTools/bfbb/compare/v0.2.1...HEAD
+[unreleased]: https://github.com/BfBBModdingTools/bfbb/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/BfBBModdingTools/bfbb/releases/tag/v0.3.0
 [0.2.1]: https://github.com/BfBBModdingTools/bfbb/releases/tag/v0.2.1
 [0.2.0]: https://github.com/BfBBModdingTools/bfbb/releases/tag/v0.2.0
 [0.1.0]: https://github.com/BfBBModdingTools/bfbb/releases/tag/v0.1.0